requirements.txt
fastapi[standard]
motor
uvicorn
python-dotenv
pydantic
